requirements.txt
openai
requests
python-dotenv
ipython
opencv-python
moviepy